added files for helper functions and config variables, added getJSON and request timeout helpers
