+++ src/js/controller.js
@@ -2,17 +2,10 @@ import * as model from './model';
 import recipeView from './views/recipeView';
 
 import 'core-js/stable';
-import 'regenerator-runtime/runtime';
 
 const recipeContainer = document.querySelector('.recipe');
 
-const timeout = function (s) {
-  return new Promise(function (_, reject) {
-    setTimeout(function () {
-      reject(new Error(`Request took too long! Timeout after ${s} second`));
-    }, s * 1000);
-  });
-};
+
 
 // https://forkify-api.herokuapp.com/v2
