+++ src/js/model.js
@@ -1,21 +1,13 @@
-export const state = {
-    recipe: {
+import { API_URL } from "./config";
+import { getJSON } from './helpers';
 
-    }
+export const state = {
+    recipe: {}
 };
 
 export const loadRecipe = async function(id) {
     try {
-
-    const res = await fetch(
-        `https://forkify-api.herokuapp.com/api/v2/recipes/${id}`
-        );
-  
-      
-  
-      const data = await res.json();
-  
-      if(!res.ok) throw new Error(`${data.message} (${res.status})`);
+      const data = await getJSON(`${API_URL}/${id}`)
   
       const {recipe} = data.data
       state.recipe = {
